Simplified GET test to not use POST request.
We want to test the GET reqeust in isolation, so it is better to not
make the test use the POST request in order to add data to the database.
If the POST request breaks, then our GET tests will probably break as
well. It will be harder to find the code that actually broke since multiple
tests will be failing.